# packages/api-types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mini Hedge Fund Desk",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/risk/{portfolio_id}/snapshot": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Get Risk Snapshot",
        "operationId": "get_risk_snapshot_api_v1_risk__portfolio_id__snapshot_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RiskSnapshot"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Risk Snapshot Api V1 Risk  Portfolio Id  Snapshot Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Take Risk Snapshot",
        "operationId": "take_risk_snapshot_api_v1_risk__portfolio_id__snapshot_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/{portfolio_id}/snapshot/history": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Get Risk History",
        "operationId": "get_risk_history_api_v1_risk__portfolio_id__snapshot_history_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskSnapshot"
                  },
                  "title": "Response Get Risk History Api V1 Risk  Portfolio Id  Snapshot History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/{portfolio_id}/var": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Calculate Var",
        "operationId": "calculate_var_api_v1_risk__portfolio_id__var_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaRRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaRResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/{portfolio_id}/stress": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Run Predefined Stress Tests",
        "operationId": "run_predefined_stress_tests_api_v1_risk__portfolio_id__stress_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StressTestResult"
                  },
                  "title": "Response Run Predefined Stress Tests Api V1 Risk  Portfolio Id  Stress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Run Custom Stress Test",
        "operationId": "run_custom_stress_test_api_v1_risk__portfolio_id__stress_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomStressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StressTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/{portfolio_id}/factors": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Get Factor Decomposition",
        "operationId": "get_factor_decomposition_api_v1_risk__portfolio_id__factors_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorDecomposition"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomStressRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "shocks": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Shocks"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "shocks"
        ],
        "title": "CustomStressRequest"
      },
...
      "FactorDecomposition": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "total_risk": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Risk"
          },
          "systematic_risk": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Systematic Risk"
          },
          "idiosyncratic_risk": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Idiosyncratic Risk"
          },
          "systematic_pct": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Systematic Pct"
          },
          "factor_exposures": {
            "items": {
              "$ref": "#/components/schemas/FactorExposure"
            },
            "type": "array",
            "title": "Factor Exposures",
            "default": []
          },
          "calculated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Calculated At"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "total_risk",
          "systematic_risk",
          "idiosyncratic_risk",
          "systematic_pct",
          "calculated_at"
        ],
        "title": "FactorDecomposition",
        "description": "Factor model decomposition of portfolio risk."
      },
      "FactorExposure": {
        "properties": {
          "factor": {
            "$ref": "#/components/schemas/RiskFactor"
          },
          "factor_name": {
            "type": "string",
            "title": "Factor Name"
          },
          "beta": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Beta"
          },
          "exposure_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Exposure Value"
          },
          "pct_of_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pct Of Total"
          }
        },
        "type": "object",
        "required": [
          "factor",
          "factor_name",
          "beta",
          "exposure_value",
          "pct_of_total"
        ],
        "title": "FactorExposure"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskFactor": {
        "type": "string",
        "enum": [
          "market",
          "sector",
          "currency",
          "idiosyncratic"
        ],
        "title": "RiskFactor"
      },
...
      "RiskSnapshot": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "nav": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Nav"
          },
          "var_95_1d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Var 95 1D"
          },
          "var_99_1d": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Var 99 1D"
          },
          "expected_shortfall_95": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Expected Shortfall 95"
          },
          "max_drawdown": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Max Drawdown"
          },
          "sharpe_ratio": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe Ratio"
          },
          "snapshot_at": {
            "type": "string",
            "format": "date-time",
            "title": "Snapshot At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "portfolio_id",
          "nav",
          "var_95_1d",
          "var_99_1d",
          "expected_shortfall_95",
          "max_drawdown",
          "snapshot_at"
        ],
        "title": "RiskSnapshot",
        "description": "Complete risk snapshot for a portfolio."
      },
...
      "StressPositionImpact": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "current_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Current Value"
          },
          "stressed_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Stressed Value"
          },
          "pnl_impact": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pnl Impact"
          },
          "pct_change": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pct Change"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "current_value",
          "stressed_value",
          "pnl_impact",
          "pct_change"
        ],
        "title": "StressPositionImpact"
      },
      "StressScenarioType": {
        "type": "string",
        "enum": [
          "predefined",
          "custom"
        ],
        "title": "StressScenarioType"
      },
      "StressTestResult": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "scenario_type": {
            "$ref": "#/components/schemas/StressScenarioType"
          },
          "total_pnl_impact": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Pnl Impact"
          },
          "total_pct_change": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Pct Change"
          },
          "position_impacts": {
            "items": {
              "$ref": "#/components/schemas/StressPositionImpact"
            },
            "type": "array",
            "title": "Position Impacts",
            "default": []
          },
          "calculated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Calculated At"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "scenario_name",
          "scenario_type",
          "total_pnl_impact",
          "total_pct_change",
          "calculated_at"
        ],
        "title": "StressTestResult",
        "description": "Result of a stress test scenario."
      },
...
      "VaRContribution": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "weight": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight"
          },
          "marginal_var": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Marginal Var"
          },
          "component_var": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Component Var"
          },
          "pct_contribution": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pct Contribution"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "weight",
          "marginal_var",
          "component_var",
          "pct_contribution"
        ],
        "title": "VaRContribution"
      },
      "VaRMethod": {
        "type": "string",
        "enum": [
          "historical",
          "parametric"
        ],
        "title": "VaRMethod"
      },
      "VaRRequest": {
        "properties": {
          "method": {
            "$ref": "#/components/schemas/VaRMethod",
            "default": "historical"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.95
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days",
            "default": 1
          }
        },
        "type": "object",
        "title": "VaRRequest"
      },
      "VaRResult": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "method": {
            "$ref": "#/components/schemas/VaRMethod"
          },
          "confidence_level": {
            "type": "number",
            "title": "Confidence Level"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "var_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Var Amount"
          },
          "var_pct": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Var Pct"
          },
          "expected_shortfall": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Expected Shortfall"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/VaRContribution"
            },
            "type": "array",
            "title": "Contributions",
            "default": []
          },
          "calculated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Calculated At"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "method",
          "confidence_level",
          "horizon_days",
          "var_amount",
          "var_pct",
          "expected_shortfall",
          "calculated_at"
        ],
        "title": "VaRResult",
        "description": "Value at Risk calculation result."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}